FUSETOOLS-3028 - fix error control decoration on creation
the problem was that the element is initially created with the wrong
parent for Camel model (the CamelFile instead of the CamelContext). It
was correctly created in xml node which explains why after a reload the
error was not displayed anymore.

I also started to refactor during investigation to gather parts related
to each Action in their respective classes instead of the
RestConfigEditor.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>